Спринт 16. Создал новые пакеты, старые удалил, предварительно все из них вытащив
