+++ app/src/main/java/com/example/playlistmaker/data/old/ITunesAPI.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.data
+package com.example.playlistmaker.data.old
 
 import retrofit2.Call
 import retrofit2.http.GET
