+++ app/src/main/java/com/example/playlistmaker/data/old/NetworkClient.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.data
+package com.example.playlistmaker.data.old
 
 interface NetworkClient {
     fun doRequest(dto: Any): Response
