+++ app/src/main/java/com/example/playlistmaker/data/old/Responce.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.data
+package com.example.playlistmaker.data.old
 
 open class Response() {
     var resultCode = 0
