+++ app/src/main/java/com/example/playlistmaker/data/old/TrackDto.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.data
+package com.example.playlistmaker.data.old
 
 data class TrackDto(
     val trackId: String,
