+++ app/src/main/java/com/example/playlistmaker/data/old/TrackMapper.kt
@@ -1,6 +1,6 @@
-package com.example.playlistmaker.data
+package com.example.playlistmaker.data.old
 
-import com.example.playlistmaker.domain.Track
+import com.example.playlistmaker.domain.old.Track
 
 class TrackMapper {
     fun mapToDomain(trackDto: TrackDto): Track {
