+++ app/src/main/java/com/example/playlistmaker/data/old/TrackSearchRequest.kt
@@ -1,3 +1,3 @@
-package com.example.playlistmaker.data
+package com.example.playlistmaker.data.old
 
 data class TrackSearchRequest(val expression: String)
