+++ app/src/main/java/com/example/playlistmaker/data/old/TrackSearchResponse.kt
@@ -1,3 +1,3 @@
-package com.example.playlistmaker.data
+package com.example.playlistmaker.data.old
 
 class TrackSearchResponse(val results: List<TrackDto>) : Response()
