--- app/src/main/java/com/example/playlistmaker/domain/TracksRepository.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.domain
-
-import com.example.playlistmaker.domain.Track
-
-interface TracksRepository {
-    fun searchTrack(expression: String): List<Track>
-}
