+++ app/src/main/java/com/example/playlistmaker/domain/old/TracksInteractor.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.domain
+package com.example.playlistmaker.domain.old
 
 interface TracksInteractor {
     fun searchTrack(expression: String, consumer: TrackConsumer)
