+++ app/src/main/java/com/example/playlistmaker/domain/old/TracksRepository.kt
@@ -1,0 +1,5 @@
+package com.example.playlistmaker.domain.old
+
+interface TracksRepository {
+    fun searchTrack(expression: String): List<Track>
+}
