+++ app/src/main/java/com/example/playlistmaker/ui/old/PlayerActivity.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.ui
+package com.example.playlistmaker.ui.old
 
 import android.media.MediaPlayer
 import android.os.Bundle
