+++ app/src/main/java/com/example/playlistmaker/ui/old/SearchTrackHistory.kt
@@ -1,8 +1,8 @@
-package com.example.playlistmaker.ui
+package com.example.playlistmaker.ui.old
 
 import android.content.SharedPreferences
 import android.util.Log
-import com.example.playlistmaker.domain.Track
+import com.example.playlistmaker.domain.old.Track
 import com.google.gson.Gson
 
 const val HISTORY_PREFS = "Playlist_Maker_History"
