+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:layout_height="fill_parent"
     android:background="?attr/colorPrimary"
     android:orientation="vertical"
-    tools:context=".ui.MainActivity">
+    tools:context=".ui.old.MainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
