+++ app/src/main/res/layout/activity_media.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.MediaActivity">
+    tools:context=".ui.old.MediaActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
