+++ app/src/main/res/layout/activity_search.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.SearchActivity">
+    tools:context=".ui.old.SearchActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
