+++ app/src/main/res/layout/activity_settings.xml
@@ -6,7 +6,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="?attr/colorSecondary"
     android:orientation="vertical"
-    tools:context=".ui.SettingsActivity">
+    tools:context=".ui.old.SettingsActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
